Do cleanup and move helpers to relay_helpers
